requirements.txt
grpcio
grpcio-tools
mediapipe
opencv-python
